添加了API，同时修改后的数据库

+++ src/main/java/com/yixun/yixun_backend/dto/FundOutDTO.java
@@ -9,7 +9,7 @@ public class FundOutDTO {
     private int fund_out_id;
 
     @JsonProperty("fund_out_amount")
-    private int fund_out_amount;
+    private double fund_out_amount;
 
     @JsonProperty("fund_out_time")
     private String fund_out_time;
